requirements.txt
matplotlib==2.1.1
numpy==1.13.3
Pillow==4.3.0
